% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/enpls-package.R
\docType{package}
\name{enpls-package}
\alias{enpls-package}
\title{Ensemble Partial Least Squares Regression}
\description{
enpls offers an algorithmic framework for measuring feature importance,
outlier detection, model applicability evaluation, and
ensemble predictive modeling with (sparse)
partial least squares regressions.
}
\details{
See the vignette via \code{vignette("enpls")}.

\tabular{ll}{ Package: \tab enpls\cr
Type: \tab Package\cr
License: \tab GPL\cr }
}
\note{
Bug reports and feature requests should be sent to
\url{https://github.com/road2stat/enpls/issues}.
}

