% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/enpls.ad.R
\name{enpls.ad.core.pred}
\alias{enpls.ad.core.pred}
\title{core prediction function for enpls.ad}
\usage{
enpls.ad.core.pred(model, testset)
}
\value{
the error vector between predicted and real response
}
\description{
core prediction function for enpls.ad
}
\keyword{internal}

