% Generated by roxygen2 (4.0.2): do not edit by hand
\name{enpls.en.core}
\alias{enpls.en.core}
\title{core function for enpls.en}
\usage{
enpls.en.core(plsdf, maxcomp)
}
\value{
the coefficients
}
\description{
select the best ncomp with cross-validation and
use it to fit the complete training set again.
scale = TRUE
}
\keyword{internal}

