% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/enpls.en.R
\name{enpls.en}
\alias{enpls.en}
\title{Ensemble Partial Least Squares Regression}
\usage{
enpls.en(x, y, maxcomp = NULL, MCtimes = 500L, method = c("mc",
  "bootstrap"), ratio = 0.8, parallel = 1L)
}
\arguments{
\item{x}{predictor matrix}

\item{y}{response vector}

\item{maxcomp}{Maximum number of components included within the models,
if not specified, default is the variable (column) numbers in x.}

\item{MCtimes}{times of Monte-Carlo}

\item{method}{\code{"mc"} or \code{"bootstrap"}. Default is \code{"mc"}.}

\item{ratio}{sample ratio used when \code{method = "mc"}}

\item{parallel}{Integer. Number of parallel processes to use.
Default is \code{1}, which means run serially.}
}
\value{
A list containing all PLS model objects.
}
\description{
This function performs ensemble partial least squares regression.
}
\details{
This function performs ensemble partial least squares regression.
}
\examples{
\donttest{data(alkanes)
x = alkanes$x
y = alkanes$y

set.seed(42)
enpls.fit = enpls.en(x, y)
print(enpls.fit)
predict(enpls.fit, newx = x)}
}
\author{
Nan Xiao <\email{road2stat@gmail.com}>
}
\references{
Dongsheng Cao, Yizeng Liang, Qingsong Xu, Yifeng Yun, and Hongdong Li.
"Toward better QSAR/QSPR modeling: simultaneous outlier detection and
variable selection using distribution of model features."
\emph{Journal of computer-aided molecular design} 25, no. 1 (2011): 67--80.
}
\seealso{
See \code{\link{enpls.fs}} for feature selection with ensemble PLS.
See \code{\link{enpls.od}} for outlier detection with ensemble PLS.
}

