% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/enpls.fit.R
\name{enpls.fit.core}
\alias{enpls.fit.core}
\title{core function for enpls.fit}
\usage{
enpls.fit.core(plsdf, maxcomp)
}
\value{
the coefficients
}
\description{
select the best ncomp with cross-validation and
use it to fit the complete training set again.
scale = TRUE
}
\keyword{internal}

