% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/enpls.fs.R
\name{enpls.fs.core}
\alias{enpls.fs.core}
\title{core function for enpls.fs}
\usage{
enpls.fs.core(plsdf, maxcomp)
}
\value{
fitted coefficients
}
\description{
select the best ncomp with cross-validation and
use it to fit the complete training set again.
scale = FALSE
}
\keyword{internal}

