% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/evaluate.R
\name{enpls.mae}
\alias{enpls.mae}
\title{Mean Absolute Error (MAE)}
\usage{
enpls.mae(yreal, ypred)
}
\arguments{
\item{yreal}{true response vector}

\item{ypred}{predicted response vector}
}
\value{
MAE
}
\description{
Mean Absolute Error (MAE)
}
\author{
Nan Xiao <\url{https://nanx.me}>
}

