% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/evaluate.R
\name{enpls.rmse}
\alias{enpls.rmse}
\title{Root Mean Squared Error (RMSE)}
\usage{
enpls.rmse(yreal, ypred)
}
\arguments{
\item{yreal}{true response vector}

\item{ypred}{predicted response vector}
}
\value{
RMSE
}
\description{
Compute Root Mean Squared Error (RMSE).
}
\author{
Nan Xiao <\url{https://nanx.me}>
}

