% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/evaluate.R
\name{enpls.rmsle}
\alias{enpls.rmsle}
\title{Root Mean Squared Logarithmic Error (RMSLE)}
\usage{
enpls.rmsle(yreal, ypred)
}
\arguments{
\item{yreal}{true response vector}

\item{ypred}{predicted response vector}
}
\value{
RMSLE
}
\description{
Root Mean Squared Logarithmic Error (RMSLE)
}
\author{
Nan Xiao <\url{https://nanx.me}>
}
