% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/enspls.ad.R
\name{enspls.ad.core.pred}
\alias{enspls.ad.core.pred}
\title{core prediction function for enspls.ad}
\usage{
enspls.ad.core.pred(model, x.te, y.te)
}
\value{
the error vector between predicted and real response
}
\description{
core prediction function for enspls.ad
}
\keyword{internal}

