% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/enspls.fit.R
\name{enspls.fit.core}
\alias{enspls.fit.core}
\title{core function for enspls.fit}
\usage{
enspls.fit.core(xtmp, ytmp, maxcomp, cvfolds, alpha)
}
\value{
the coefficients
}
\description{
select the best ncomp and alpha, then use them to fit
the complete training set.
scale = TRUE
}
\keyword{internal}
