% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/enspls.fs.R
\name{enspls.fs.core}
\alias{enspls.fs.core}
\title{core function for enspls.fs}
\usage{
enspls.fs.core(xtmp, ytmp, maxcomp, cvfolds, alpha)
}
\value{
fitted coefficients
}
\description{
select the best ncomp and alpha with cross-validation, then
use it to fit the complete training set.
scale = FALSE
}
\keyword{internal}
