% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/enpls-datalist.R
\docType{data}
\name{logd1k}
\alias{logd1k}
\title{logD7.4 Data for 1,000 Compounds}
\format{A list with 2 components:
\itemize{
\item x - data frame with 1,000 rows (samples) and 80 columns (predictors)
\item y - numeric vector of length 1,000 (response)
}
The first 1000 compounds in the original dataset were selected.}
\usage{
data(logd1k)
}
\description{
Distribution coefficients at pH 7.4 (logD7.4) dataset from Wang et, al.
}
\details{
This dataset contains distribution coefficients at pH 7.4 (logD7.4)
for 1,000 compounds, and their 80 molecular descriptors computed with RDKit.
}
\examples{
data(logd1k)
str(logd1k)
}
\references{
Jian-Bing Wang, Dong-Sheng Cao, Min-Feng Zhu, Yong-Huan Yun, Nan Xiao,
and Yi-Zeng Liang. "In silico evaluation of logD7.4 and comparison with
other prediction methods."
\emph{Journal of Chemometrics} 29, no. 7 (2015): 389--398.
}

