% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot.cv.enpls}
\alias{plot.cv.enpls}
\title{Plot cv.enpls object}
\usage{
\method{plot}{cv.enpls}(x, main = NULL, ...)
}
\arguments{
\item{x}{An object of class \code{cv.enpls}.}

\item{main}{plot title}

\item{...}{Additional graphical parameters for \code{\link{plot}}.}
}
\description{
Plot cv.enpls object
}
\examples{
data(alkanes)
x = alkanes$x
y = alkanes$y

set.seed(42)
cv.enpls.fit = cv.enpls(x, y, MCtimes = 20)
plot(cv.enpls.fit)
}
\author{
Nan Xiao <\url{http://nanx.me}>
}
\seealso{
See \code{\link{cv.enpls}} for performing ensemble PLS regression.
}

