% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.enspls.R
\name{plot.cv.enspls}
\alias{plot.cv.enspls}
\title{Plot cv.enspls object}
\usage{
\method{plot}{cv.enspls}(x, xlim = NULL, ylim = NULL, main = NULL, ...)
}
\arguments{
\item{x}{An object of class \code{cv.enspls}.}

\item{xlim}{x vector of lenght 2, x axis limits of the plot}

\item{ylim}{y vector of lenght 2, y axis limits of the plot}

\item{main}{Plot title, not used.}

\item{...}{Additional graphical parameters, not used.}
}
\description{
Plot cv.enspls object
}
\examples{
# This example takes one minute to run
\dontrun{
data("logd1k")
x = logd1k$x
y = logd1k$y

set.seed(42)
cvfit = cv.enspls(x, y, reptimes = 10)
plot(cvfit)}
}
\author{
Nan Xiao <\url{http://nanx.me}>
}
\seealso{
See \code{\link{cv.enspls}} for ensemble sparse
partial least squares regression.
}

