% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.enpls.R
\name{plot.enpls.ad}
\alias{plot.enpls.ad}
\title{Plot enpls.ad object}
\usage{
\method{plot}{enpls.ad}(x, type = c("static", "interactive"), main = NULL,
  ...)
}
\arguments{
\item{x}{An object of class \code{enpls.ad}.}

\item{type}{Plot type. Can be \code{"static"} or \code{"interactive"}.}

\item{main}{Plot title, not used currently.}

\item{...}{Additional graphical parameters, not used currently.}
}
\description{
Plot enpls.ad object
}
\examples{
data("alkanes")

# training set
x = alkanes$x[1:100, ]
y = alkanes$y[1:100]

# two test sets
xtest = list("test1" = alkanes$x[101:150, ],
             "test2" = alkanes$x[151:207, ])
ytest = list("test1" = alkanes$y[101:150],
             "test2" = alkanes$y[151:207])

set.seed(42)
ad = enpls.ad(x, y, xtest, ytest,
              space = "variable", method = "mc",
              ratio = 0.8, reptimes = 50)

plot(ad)
# The interactive plot requires a HTML viewer
\dontrun{
plot(ad, type = "interactive")}
}
\author{
Nan Xiao <\url{http://nanx.me}>
}
\seealso{
See \code{\link{enpls.ad}} for model applicability domain
evaluation with ensemble partial least squares regressions.
}

