% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/plot.R
\name{plot.enpls.fs}
\alias{plot.enpls.fs}
\title{Plot enpls.fs Object}
\usage{
\method{plot}{enpls.fs}(x, sort = TRUE, nvar = NULL, main = NULL, ...)
}
\arguments{
\item{x}{An object of class \code{enpls.fs}.}

\item{sort}{Should the variables be sorted in decreasing order of importance?}

\item{nvar}{How many variables to show? Ignored if \code{sort = FALSE}.}

\item{main}{plot title}

\item{...}{Other graphical parameters to be passed on to \code{dotchart}.}
}
\description{
This function plots enpls.fs object.
}
\details{
This function plots enpls.fs object.
}
\examples{
\donttest{
data(alkanes)
x = alkanes$x
y = alkanes$y

set.seed(42)
varimp = enpls.fs(x, y)
plot(varimp)
plot(varimp, nvar = 10L)}
}
\author{
Nan Xiao <\email{road2stat@gmail.com}>
}
\seealso{
See \code{\link{enpls.fs}} for feature selection with ensemble PLS.
}

