% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict.R
\name{predict.enpls.fit}
\alias{predict.enpls.fit}
\title{Make Predictions from a Fitted Ensemble Partial Least Squares Model}
\usage{
\method{predict}{enpls.fit}(object, newx, method = c("mean", "median"), ...)
}
\arguments{
\item{object}{An object of class \code{enpls.fit}.}

\item{newx}{New data to predict with.}

\item{method}{Use \code{mean} or \code{median} as the final prediction.}

\item{...}{Additional parameters for \code{\link{predict}}.}
}
\value{
A numeric vector containing the predicted values.
}
\description{
Make predictions on new data by fitted enpls.fit object.
}
\details{
Make predictions on new data by fitted enpls.fit object.
}
\examples{
data("alkanes")
x = alkanes$x
y = alkanes$y

set.seed(42)
fit = enpls.fit(x, y, MCtimes = 100)
y.pred = predict(fit, newx = x)
plot(y, y.pred, xlim = range(y), ylim = range(y))
abline(a = 0L, b = 1L)
y.pred.med = predict(fit, newx = x, method = 'median')
plot(y, y.pred.med, xlim = range(y), ylim = range(y))
abline(a = 0L, b = 1L)
}
\author{
Nan Xiao <\url{http://nanx.me}>
}
\references{
Dongsheng Cao, Yizeng Liang, Qingsong Xu, Yifeng Yun, and Hongdong Li.
"Toward better QSAR/QSPR modeling: simultaneous outlier detection and
variable selection using distribution of model features."
\emph{Journal of computer-aided molecular design} 25, no. 1 (2011): 67--80.
}
\seealso{
See \code{\link{enpls.fit}} for ensemble
partial least squares regression.
}

