% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.enpls.R
\name{print.enpls.ad}
\alias{print.enpls.ad}
\title{Print enpls.ad Object}
\usage{
\method{print}{enpls.ad}(x, ...)
}
\arguments{
\item{x}{An object of class \code{enpls.ad}.}

\item{...}{Additional parameters for \code{\link{print}}.}
}
\description{
Print enpls.ad object.
}
\examples{
data("alkanes")

# training set
x = alkanes$x[1:100, ]
y = alkanes$y[1:100]

# two test sets
xtest = list("test1" = alkanes$x[101:150, ],
             "test2" = alkanes$x[151:207, ])
ytest = list("test1" = alkanes$y[101:150],
             "test2" = alkanes$y[151:207])

set.seed(42)
ad = enpls.ad(x, y, xtest, ytest,
              space = "variable", method = "mc",
              ratio = 0.8, reptimes = 50)
print(ad)
}
\author{
Nan Xiao <\url{http://nanx.me}>
}
\seealso{
See \code{\link{enpls.ad}} for model applicability domain
evaluation with ensemble partial least squares regressions.
}

