% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.R
\name{print.enpls.en}
\alias{print.enpls.en}
\title{Print Fitted Ensemble Partial Least Squares Object}
\usage{
\method{print}{enpls.en}(x, ...)
}
\arguments{
\item{x}{An object of class \code{enpls.en}.}

\item{...}{Additional parameters for \code{\link{print}}.}
}
\description{
This function prints coefficients of each model in the enpls.en object.
}
\details{
This function prints coefficients of each model in the enpls.en object.
}
\examples{
data(alkanes)
x = alkanes$x
y = alkanes$y

set.seed(42)
enpls.fit = enpls.en(x, y, MCtimes = 100)
print(enpls.fit)
}
\author{
Nan Xiao <\url{http://nanx.me}>
}
\seealso{
See \code{\link{enpls.en}} for performing ensemble PLS regression.
}

