% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.enpls.R
\name{print.enpls.fit}
\alias{print.enpls.fit}
\title{Print Fitted Ensemble Partial Least Squares Object}
\usage{
\method{print}{enpls.fit}(x, ...)
}
\arguments{
\item{x}{An object of class \code{enpls.fit}.}

\item{...}{Additional parameters for \code{\link{print}}.}
}
\description{
Print coefficients of each model in the enpls.fit object.
}
\examples{
data("alkanes")
x = alkanes$x
y = alkanes$y

set.seed(42)
fit = enpls.fit(x, y, reptimes = 50)
print(fit)
}
\author{
Nan Xiao <\url{http://nanx.me}>
}
\seealso{
See \code{\link{enpls.fit}} for ensemble
partial least squares regression.
}

