% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/print.R
\name{print.enpls.od}
\alias{print.enpls.od}
\title{Print enpls.od Object}
\usage{
\method{print}{enpls.od}(x, ...)
}
\arguments{
\item{x}{An object of class \code{enpls.od}.}

\item{...}{Other parameters to be passed on to \code{print}.}
}
\description{
This function prints enpls.od object.
}
\details{
This function prints enpls.od object.
}
\examples{
\donttest{
data(alkanes)
x = alkanes$x
y = alkanes$y

set.seed(42)
od = enpls.od(x, y)
print(od)}
}
\author{
Nan Xiao <\email{road2stat@gmail.com}>
}
\seealso{
See \code{\link{enpls.od}} for outlier detection with ensemble PLS.
}

