% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.enspls.R
\name{print.enspls.od}
\alias{print.enspls.od}
\title{Print enspls.od Object}
\usage{
\method{print}{enspls.od}(x, ...)
}
\arguments{
\item{x}{An object of class \code{enspls.od}.}

\item{...}{Additional parameters for \code{\link{print}}.}
}
\description{
Print enspls.od object.
}
\examples{
data("logd1k")
x = logd1k$x
y = logd1k$y

set.seed(42)
od = enspls.od(x, y, reptimes = 5, maxcomp = 3,
               alpha = c(0.3, 0.6, 0.9))
print(od)
}
\author{
Nan Xiao <\url{https://nanx.me}>
}
\seealso{
See \code{\link{enspls.od}} for outlier detection with
ensemble sparse partial least squares regressions.
}

