% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.enpls.R
\name{rgb2alpha}
\alias{rgb2alpha}
\title{Add transparency level to hex colors}
\usage{
rgb2alpha(raw_col, alpha)
}
\value{
Color hex value (with transparency)
}
\description{
Add transparency level to hex colors
}
\keyword{internal}
