% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cv.enspls.R
\name{cv.enspls}
\alias{cv.enspls}
\title{Cross Validation for Ensemble Sparse Partial Least Squares Regression}
\usage{
cv.enspls(x, y, nfolds = 5L, verbose = TRUE, ...)
}
\arguments{
\item{x}{Predictor matrix.}

\item{y}{Response vector.}

\item{nfolds}{Number of cross-validation folds, default is \code{5}.
Note that this is the CV folds for the ensemble sparse PLS model,
not the individual sparse PLS models. To control the CV folds for
single sparse PLS models, please use the argument \code{cvfolds}.}

\item{verbose}{Shall we print out the progress of cross-validation?}

\item{...}{Arguments to be passed to \code{\link{enspls.fit}}.}
}
\value{
A list containing:
\itemize{
\item \code{ypred} - a matrix containing two columns: real y and predicted y
\item \code{residual} - cross validation result (y.pred - y.real)
\item \code{RMSE} - RMSE
\item \code{MAE} - MAE
\item \code{Rsquare} - Rsquare
}
}
\description{
K-fold cross validation for ensemble sparse partial least squares regression.
}
\note{
To maximize the probablity that each observation can
be selected in the test set (thus the prediction uncertainty
can be measured), please try setting a large \code{reptimes}.
}
\examples{
# This example takes one minute to run
\dontrun{
data("logd1k")
x = logd1k$x
y = logd1k$y

set.seed(42)
cvfit = cv.enspls(x, y, reptimes = 10)
print(cvfit)
plot(cvfit)}
}
\author{
Nan Xiao <\url{https://nanx.me}>
}
\seealso{
See \code{\link{enspls.fit}} for ensemble sparse
partial least squares regressions.
}

