% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/enpls-package.R
\docType{package}
\name{enpls-package}
\alias{enpls-package}
\title{Ensemble Partial Least Squares Regression}
\description{
The enpls package provides functions for running the 
ensemble partial least squares regression.
}
\details{
The comprehensive user guide can be opened with 
\code{vignette('enpls')}.

\tabular{ll}{ Package: \tab enpls\cr Type: \tab Package\cr
Version: \tab 2.0\cr License: \tab GPL (>= 2)\cr }
}
\note{
Bug reports and feature requests should be sent to
\url{https://github.com/road2stat/enpls/issues}.
}

