% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/enpls-package.R
\docType{package}
\name{enpls-package}
\alias{enpls-package}
\title{Ensemble (Sparse) Partial Least Squares Regressions}
\description{
An algorithmic framework for measuring feature importance,
outlier detection, model applicability evaluation, and
ensemble predictive modeling with (sparse)
partial least squares regressions.
}
\details{
See the vignette via \code{vignette("enpls")}.

\tabular{ll}{ Package: \tab enpls\cr
Type: \tab Package\cr
License: \tab GPL-3\cr }
}

