% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/enpls.R
\docType{package}
\name{enpls-package}
\alias{enpls}
\alias{enpls-package}
\title{enpls: Ensemble Partial Least Squares Regression}
\description{
An algorithmic framework for measuring feature importance,
    outlier detection, model applicability domain evaluation,
    and ensemble predictive modeling with (sparse)
    partial least squares regressions.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://nanx.me/enpls/}
  \item \url{https://github.com/nanxstats/enpls}
  \item Report bugs at \url{https://github.com/nanxstats/enpls/issues}
}

}
\author{
\strong{Maintainer}: Nan Xiao \email{me@nanx.me} (\href{https://orcid.org/0000-0002-0250-5673}{ORCID})

Authors:
\itemize{
  \item Dong-Sheng Cao
  \item Miao-Zhu Li \email{miaozhu.li@duke.edu}
  \item Qing-Song Xu \email{qsxu@csu.edu.cn}
}

}
\keyword{internal}
