% Generated by roxygen2 (4.0.2): do not edit by hand
\docType{package}
\name{enpls-package}
\alias{enpls-package}
\title{Ensemble Partial Least Squares (EnPLS) Regression}
\description{
The enpls package provides functions for running the
ensemble partial least squares (EnPLS) regression.
}
\details{
The comprehensive user guide could be opened with
\code{vignette('enpls')}.

The enpls package is developed by Computational Biology and
Drug Design (CBDD) Group, Central South University
(\url{http://cbdd.csu.edu.cn/}).

\tabular{ll}{ Package: \tab enpls\cr Type: \tab Package\cr
Version: \tab 1.0\cr License: \tab GPL (>= 2)\cr }
}
\note{
Bug reports and feature requests should be sent to
\url{https://github.com/road2stat/enpls/issues}.
}
\author{
Nan Xiao <\email{road2stat@gmail.com}>
        Dongsheng Cao <\email{oriental-cds@163.com}>
        Qingsong Xu <\email{dasongxu@gmail.com}>
}

