% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/enpls.fs.R
\name{enpls.fs}
\alias{enpls.fs}
\title{Ensemble Partial Least Squares for Feature Selection}
\usage{
enpls.fs(x, y, maxcomp = NULL, reptimes = 500L, method = c("mc",
  "bootstrap"), ratio = 0.8, parallel = 1L)
}
\arguments{
\item{x}{predictor matrix}

\item{y}{response vector}

\item{maxcomp}{Maximum number of components included within the models,
if not specified, default is the variable (column) numbers in x.}

\item{reptimes}{Number of models to build with Monte-Carlo resampling
or bootstrapping.}

\item{method}{\code{"mc"} or \code{"bootstrap"}. Default is \code{"mc"}.}

\item{ratio}{sample ratio used when \code{method = "mc"}}

\item{parallel}{Integer. Number of CPU cores to use.
Default is \code{1} (not parallelized).}
}
\value{
A list containing two components:
\itemize{
\item \code{variable.importance} - a vector of variable importance
\item \code{coefficient.matrix} - original coefficient matrix
}
}
\description{
Feature selection with ensemble partial least squares.
}
\examples{
data("alkanes")
x = alkanes$x
y = alkanes$y

set.seed(42)
fs = enpls.fs(x, y, reptimes = 50)
print(fs)
plot(fs)
}
\author{
Nan Xiao <\url{http://nanx.me}>
}
\seealso{
See \code{\link{enpls.od}} for outlier detection with
ensemble partial least squares regression.
See \code{\link{enpls.fit}} for ensemble partial least squares regression.
}

