% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/plot.R
\name{plot.cv.enpls}
\alias{plot.cv.enpls}
\title{Plot cv.enpls Object}
\usage{
\method{plot}{cv.enpls}(x, main = NULL, ...)
}
\arguments{
\item{x}{An object of class \code{cv.enpls}.}

\item{main}{plot title}

\item{...}{Other graphical parameters to be passed on to \code{plot}.}
}
\description{
This function plots cv.enpls object.
}
\details{
This function plots cv.enpls object.
}
\examples{
\donttest{
data(alkanes)
x = alkanes$x
y = alkanes$y

set.seed(42)
cv.enpls.fit = cv.enpls(x, y)
plot(cv.enpls.fit)}
}
\author{
Nan Xiao <\email{road2stat@gmail.com}>
}
\seealso{
See \code{\link{cv.enpls}} for performing ensemble PLS regression.
}

