% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.enpls.R
\name{plot.cv.enpls}
\alias{plot.cv.enpls}
\title{Plot cv.enpls object}
\usage{
\method{plot}{cv.enpls}(x, xlim = NULL, ylim = NULL, main = NULL, ...)
}
\arguments{
\item{x}{An object of class \code{cv.enpls}.}

\item{xlim}{x vector of lenght 2, x axis limits of the plot}

\item{ylim}{y vector of lenght 2, y axis limits of the plot}

\item{main}{Plot title, not used.}

\item{...}{Additional graphical parameters, not used.}
}
\description{
Plot cv.enpls object
}
\examples{
data("alkanes")
x = alkanes$x
y = alkanes$y

set.seed(42)
cvfit = cv.enpls(x, y, MCtimes = 10)
plot(cvfit)
}
\author{
Nan Xiao <\url{http://nanx.me}>
}
\seealso{
See \code{\link{cv.enpls}} for ensemble
partial least squares regression.
}

