% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot.enpls.od}
\alias{plot.enpls.od}
\title{Plot enpls.od object}
\usage{
\method{plot}{enpls.od}(x, criterion = c("quantile", "sd"), prob = 0.05,
  sdtimes = 3L, main = NULL, ...)
}
\arguments{
\item{x}{An object of class \code{enpls.od}.}

\item{criterion}{Criterion of being outlier, 
could be \code{'quantile'} or \code{'sd'}.}

\item{prob}{the quantile}

\item{sdtimes}{the times of sd}

\item{main}{plot title}

\item{...}{Additional graphical parameters for \code{\link{plot}}.}
}
\description{
Plot enpls.od object
}
\examples{
data(alkanes)
x = alkanes$x
y = alkanes$y

set.seed(42)
od = enpls.od(x, y, MCtimes = 100)
plot(od, criterion = 'quantile')
plot(od, criterion = 'sd')
}
\author{
Nan Xiao <\url{http://nanx.me}>
}
\seealso{
See \code{\link{enpls.od}} for outlier detection with ensemble PLS.
}

