% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.R
\name{print.cv.enpls}
\alias{print.cv.enpls}
\title{Print cv.enpls Object}
\usage{
\method{print}{cv.enpls}(x, ...)
}
\arguments{
\item{x}{An object of class \code{cv.enpls}.}

\item{...}{Additional parameters for \code{\link{print}}.}
}
\description{
Print cv.enpls object.
}
\details{
Print cv.enpls object.
}
\examples{
data("alkanes")
x = alkanes$x
y = alkanes$y

set.seed(42)
cvfit = cv.enpls(x, y, MCtimes = 20)
print(cvfit)
}
\author{
Nan Xiao <\url{http://nanx.me}>
}
\seealso{
See \code{\link{cv.enpls}} for ensemble
partial least squares regression.
}

