% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.R
\name{print.cv.enpls}
\alias{print.cv.enpls}
\title{Print cv.enpls Object}
\usage{
\method{print}{cv.enpls}(x, ...)
}
\arguments{
\item{x}{An object of class \code{cv.enpls}.}

\item{...}{Other parameters to be passed on to \code{print}.}
}
\description{
This function prints cv.enpls object.
}
\details{
This function prints cv.enpls object.
}
\examples{
data(alkanes)
x = alkanes$x
y = alkanes$y

set.seed(42)
cv.enpls.fit = cv.enpls(x, y, MCtimes = 20)
print(cv.enpls.fit)
}
\author{
Nan Xiao <\email{road2stat@gmail.com}>
}
\seealso{
See \code{\link{cv.enpls}} for performing ensemble PLS regression.
}

