% Generated by roxygen2 (4.0.2): do not edit by hand
\name{print.enpls.en}
\alias{print.enpls.en}
\title{Print Fitted Ensemble Partial Least Squares Object}
\usage{
\method{print}{enpls.en}(x, ...)
}
\arguments{
\item{x}{An object of class \code{enpls.en}.}

\item{...}{Other parameters to be passed on to \code{print}.}
}
\description{
This function prints coefficients of each model in the enpls.en object.
}
\details{
This function prints coefficients of each model in the enpls.en object.
}
\examples{
\donttest{
data(alkanes)
x = alkanes$x
y = alkanes$y

set.seed(42)
enpls.fit = enpls.en(x, y)
print(enpls.fit)}
}
\author{
Nan Xiao <\email{road2stat@gmail.com}>
}
\seealso{
See \code{\link{enpls.en}} for performing ensemble PLS regression.
}

