% Generated by roxygen2 (4.0.2): do not edit by hand
\name{print.enpls.fs}
\alias{print.enpls.fs}
\title{Print enpls.fs Object}
\usage{
\method{print}{enpls.fs}(x, sort = TRUE, nvar = NULL, ...)
}
\arguments{
\item{x}{An object of class \code{enpls.fs}.}

\item{sort}{Should the variables be sorted in decreasing order of importance?}

\item{nvar}{How many variables to show? Ignored if \code{sort = FALSE}.}

\item{...}{Other parameters to be passed on to \code{print}.}
}
\description{
This function prints enpls.fs object.
}
\details{
This function prints enpls.fs object.
}
\examples{
\donttest{
data(alkanes)
x = alkanes$x
y = alkanes$y

set.seed(42)
varimp = enpls.fs(x, y)
print(varimp)
print(varimp, nvar = 10L)}
}
\author{
Nan Xiao <\email{road2stat@gmail.com}>
}
\seealso{
See \code{\link{enpls.fs}} for feature selection with ensemble PLS.
}

