% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.enpls.R
\name{print.enpls.od}
\alias{print.enpls.od}
\title{Print enpls.od Object}
\usage{
\method{print}{enpls.od}(x, ...)
}
\arguments{
\item{x}{An object of class \code{enpls.od}.}

\item{...}{Additional parameters for \code{\link{print}}.}
}
\description{
Print enpls.od object.
}
\examples{
data("alkanes")
x = alkanes$x
y = alkanes$y

set.seed(42)
od = enpls.od(x, y, MCtimes = 50)
print(od)
}
\author{
Nan Xiao <\url{http://nanx.me}>
}
\seealso{
See \code{\link{enpls.od}} for outlier detection with
ensemble partial least squares regression.
}

