% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.enspls.R
\name{print.enspls.ad}
\alias{print.enspls.ad}
\title{Print enspls.ad Object}
\usage{
\method{print}{enspls.ad}(x, ...)
}
\arguments{
\item{x}{An object of class \code{enspls.ad}.}

\item{...}{Additional parameters for \code{\link{print}}.}
}
\description{
Print enspls.ad object.
}
\examples{
data("logd1k")

# training set
x = logd1k$x[1:200, ]
y = logd1k$y[1:200]

# two test sets
xtest = list("test1" = logd1k$x[201:300, ],
             "test2" = logd1k$x[301:400, ])
ytest = list("test1" = logd1k$y[201:300],
             "test2" = logd1k$y[301:400])

set.seed(42)
ad = enspls.ad(x, y, xtest, ytest,
               maxcomp = 3, alpha = c(0.3, 0.6, 0.9),
               space = "variable", method = "mc",
               ratio = 0.8, reptimes = 20)
print(ad)
}
\author{
Nan Xiao <\url{http://nanx.me}>
}
\seealso{
See \code{\link{enspls.ad}} for model applicability domain
evaluation with ensemble sparse partial least squares regressions.
}

